scripts/python/requirements.txt
appdirs==1.4.4
attrs==20.3.0
broadlink==0.17.0
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
future==0.18.2
importlib-metadata==2.0.0
importlib-resources==3.3.0
lxml==4.6.2
more-itertools==8.6.0
packaging==20.8
pexpect==4.8.0
pluggy==0.13.1
prettytable==0.7.2
py==1.10.0
pyaes==1.6.1
pyparsing==2.4.7
pyserial==3.5
pytest-html==2.1.1
regex==2020.11.13
requests==2.24.0
scipy==1.5.4
scp==0.13.2
six==1.15.0
toml==0.10.2
tox==3.20.1
virtualenv==20.2.2
uiautomator2==2.14.1
pandas==1.1.5
pygal==2.4.0
openpyxl==3.0.5
xlrd==1.2.0
xlwt==1.3.0
numpy==1.22.4
XlsxWriter==1.3.7
pytest-rerunfailures==11.1.2
pytest-repeat==0.9.1
pytest-dependency==0.5.1
threadpool==1.3.2
selenium
matplotlib
pyyaml
adbutils==0.14.1
uiautomator
xlutils==2.0.0
pykeyboard
PyUserInput
allure-pytest
pymysql
